repl.c: read from files as one block
This commit makes the REPL reading functions use the isatty() function
to detect whether the file will be read from keyboard or from a pipe
(such as < hello.pas).

There was a bug when reading via pipe that caused empty lines inside a
file to trigger a TOK_EOF token, causing multiple "files" to be read
separately. After this fix, if the isatty() function reveals that it is
being read from a pipe, it reads the entire file at once, ignoring empty
lines.